+++ .idea/MypyPlugin.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="MypySettings">
+    <option name="arguments" value="--follow-imports silent --exclude \.pyi$" />
+    <option name="mypyExecutable" value="$USER_HOME$/miniconda3/bin/mypy" />
+    <option name="projectDirectory" value="$PROJECT_DIR$" />
+  </component>
+</project>
