+++ src/meddra_graph/tests/test_data/smq_list.asc
