+++ src/meddra_graph/tests/test_data/soc.asc
