+++ src/meddra_graph/tests/test_data/soc_hlgt.asc
